requirements.txt
Package              Version            
-------------------- -------------------
absl-py              0.8.1              
asgiref              3.2.7              
astor                0.8.0              
certifi              2019.11.28         
Django               3.0.5              
gast                 0.2.2              
google-pasta         0.1.8              
grpcio               1.16.1             
h5py                 2.8.0              
image                1.5.31             
Keras-Applications   1.0.8              
Keras-Preprocessing  1.1.0              
Markdown             3.1.1              
mkl-fft              1.0.15             
mkl-random           1.1.0              
mkl-service          2.3.0              
numpy                1.17.4             
opt-einsum           3.1.0              
Pillow               7.1.1              
pip                  19.3.1             
protobuf             3.11.2             
pytz                 2019.3             
scipy                1.3.2              
setuptools           42.0.2.post20191203
six                  1.13.0             
sqlparse             0.3.1              
tensorboard          2.0.0              
tensorflow           2.0.0              
tensorflow-estimator 2.0.0              
termcolor            1.1.0              
Werkzeug             0.16.0             
wheel                0.33.6             
wincertstore         0.2                
wrapt                1.11.2             
